requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
httpx==0.27.0
python-dotenv==1.0.1
apscheduler==3.10.4
py-clob-client==0.16.0
Jinja2==3.1.4
